refact: move queue resolver in right folder (#71)
